Добавить возврат к статье после отправки коментария, доработать счетчики лайков и просмотров

+++ database/migrations/2022_02_08_092515_create_articles_table.php
@@ -18,8 +18,8 @@ class CreateArticlesTable extends Migration
             $table->timestamps();
             $table->string('title');
             $table->text('content');
-            $table->integer('likes');
-            $table->integer('views');
+            $table->unsignedBigInteger('likes');
+            $table->unsignedBigInteger('views');
         });
     }
